Instrumented and added assertions for Eventhubs samples
