+++ azure-samples/src/main/java/com/azure/android/MainActivity.java
@@ -3,19 +3,12 @@ package com.azure.android;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
-import android.util.Log;
 
 import com.azure.android.appconfiguration.HelloWorld;
 import com.azure.android.appconfiguration.SecretReferenceConfigurationSettingSample;
 import com.azure.android.appconfiguration.WatchFeature;
 import com.azure.android.appconfiguration.ConditionalRequestAsync;
 
-import com.azure.android.eventhubs.ConsumeEvents;
-import com.azure.android.eventhubs.EventProcessorClientAggregateEventsSample;
-import com.azure.android.eventhubs.EventProcessorClientCheckpointing;
-import com.azure.android.eventhubs.EventProcessorClientSample;
-import com.azure.android.eventhubs.PublishEventsBufferedProducer;
-import com.azure.android.eventhubs.PublishEventsWithAzureIdentity;
 import com.azure.android.keyvault.certificates.HelloWorldKeyvaultCerificates;
 import com.azure.android.keyvault.certificates.ListOperationsKeyvaultCerificates;
 import com.azure.android.keyvault.certificates.ManagingDeletedCertificatesAsyncKeyvaultCerificates;
@@ -29,8 +22,6 @@ import com.azure.android.keyvault.secrets.ManagingDeletedSecretsKeyvaultSecrets;
 import com.azure.android.storage.BasicExample;
 import com.azure.identity.ClientSecretCredential;
 import com.azure.identity.ClientSecretCredentialBuilder;
-import com.azure.identity.DefaultAzureCredential;
-import com.azure.identity.DefaultAzureCredentialBuilder;
 
 import java.io.IOException;
 
@@ -50,8 +41,6 @@ public class MainActivity extends AppCompatActivity {
         final String keyvaultEndpoint = "https://android-key-vault.vault.azure.net/";
         final String appconfigEndpoint = "https://android-app-configuration.azconfig.io";
         final String storageAccountName = "androidazsdkstorage";
-        final String eventhubsNamespace= "android-eventhubs.servicebus.windows.net";
-        final String eventhubsName = "android-eh-instance";
         Thread thread = new Thread(() -> {
             try {
                 //appconfig sample block
@@ -74,7 +63,6 @@ public class MainActivity extends AppCompatActivity {
                 }
 
                 KeyRotationAsyncKeyvaultKeys.main(keyvaultEndpoint, clientSecretCredential);
-
 
                 /* commented out pending key-id being obtained to put in this
                 try {
@@ -124,24 +112,6 @@ public class MainActivity extends AppCompatActivity {
                 } catch (IOException e) {
                     throw new RuntimeException(e);
                 }
-
-                // Eventhubs sample block
-                String[] eventargs = {eventhubsNamespace, eventhubsName};
-                Log.i("Eventhubs", "Starting ConsumeEvents");
-                ConsumeEvents.main(eventargs, clientSecretCredential);
-                Log.i("Eventhubs", "Starting PublishEvents");
-                PublishEventsWithAzureIdentity.main(eventargs, clientSecretCredential);
-                Log.i("Eventhubs", "Starting PublishEventsBuffered");
-                PublishEventsBufferedProducer.main(eventargs, clientSecretCredential);
-                
-                //TODO: Need to determine ways to verify that these three are working as intended.
-                //The second two are currently not in the app.
-                //Log.i("Eventhubs", "Starting ClientSample");
-                //EventProcessorClientSample.main(eventargs, clientSecretCredential);
-                /*Log.i("Eventhubs", "Starting ClientSample2");
-                EventProcessorClientAggregateEventsSample.main(eventargs, clientSecretCredential);*/
-                //Log.i("Eventhubs", "Starting ProcessorClient2");
-                //EventProcessorClientCheckpointing.main(eventargs, clientSecretCredential);
                 finish();
             } catch (Exception e) {
                 e.printStackTrace();
